# server/api/docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "API dos eventos de frontend",
    "version": "1.0.0",
    "title": "Frontend BR - Eventos",
    "contact": {
      "email": "angeliski@hotmail.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://github.com/frontendbr/eventos/blob/develop/LICENSE.md"
    }
  },
  "host": "frontendbr.com.br",
  "basePath": "/eventos/api/v1",
...
  "schemes": [
    "http"
  ],
  "paths": {
    "/event": {
      "post": {
        "tags": [
          "event"
        ],
        "summary": "Add a new event",
        "description": "",
        "operationId": "addEvent",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Event object that needs to be added",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Event"
            }
          }
        ],
        "responses": {
          "405": {
            "description": "Invalid input"
          }
        }
      },
      "get": {
        "tags": [
          "event"
        ],
        "summary": "List all events",
        "description": "",
        "operationId": "listEvent",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "month",
            "type": "integer",
            "minimum": 1,
            "maximum": 12,
            "required": false,
            "description": "The month."
          },
          {
            "in": "query",
            "name": "state",
            "type": "string",
            "required": false,
            "description": "The state."
          },
          {
            "in": "query",
            "name": "keywords",
            "type": "string",
            "required": false,
            "description": "keywords for search."
          },
          {
            "in": "query",
            "name": "offset",
            "type": "integer",
            "required": false,
            "default": 0,
            "minimum": 0,
            "description": "The number of items to skip before starting to collect the result set."
          },
          {
            "in": "query",
            "name": "limit",
            "type": "integer",
            "required": false,
            "default": 20,
            "minimum": 1,
            "maximum": 100,
            "description": "The numbers of items to return."
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Event"
              }
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "Event": {
      "type": "object",
      "required": [
        "title",
        "date",
        "local",
        "shortDescription",
        "price",
        "eventLink"
      ],
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "title": {
          "type": "string",
          "example": "Front In Sampa"
        },
        "date": {
          "type": "string",
          "format": "date",
          "example": "20/02/2017"
        },
        "local": {
          "type": "string",
          "example": "Avenida Paulista, 1000"
        },
        "shortDescription": {
          "type": "string",
          "example": "Evento para frontends discutirem sobre assuntos bacanas"
        },
        "price": {
          "type": "integer",
          "example": "100.00"
        },
        "eventLink": {
          "type": "string",
          "example": "http://link-para-o-evento.com.br"
        }
      }
    },
...
  },
...
}